Rename src directory to bomp, start adding support for console and package use
