+++ bomp/data/__init__.py
@@ -1,5 +1,5 @@
 __all__=['async_requests','purpleair_data_retriever','purpleair_outlier_remover','purpleair_raw_cleaner',
 'purpleair_thingspeak','tceq_data_retriever','tceq_interim_processor','tceq_raw_cleaner']
 
-from src.data import (async_requests, purpleair_data_retriever, purpleair_outlier_remover, purpleair_raw_cleaner)
+from bomp.data import (async_requests, purpleair_data_retriever, purpleair_outlier_remover, purpleair_raw_cleaner)
 #tceq_data_retriever,tceq_interim_processor,tceq_raw_cleaner)
