+++ setup.py
@@ -5,7 +5,7 @@ setup(
     version='0.1.0',
     author='Calvin J Lin',
     author_email='calvin.lin@utexas.edu',
-    packages=['src'],
-    entry_points = {'console_scripts':['bomp=src.__main__:main']},
+    packages=['bomp'],
+    entry_points = {'console_scripts':['bomp=bomp.__main__:main']},
     description='Package that can download data from UT\'s PurpleAir netork'
 )
